shelve: deleting chunks from patches mostly works
Currently no way to delete a file when it has no chunks, after asking for its
header to be kept.  And the context menu in general for the file list is mostly
innapropriate.  The same is true of revdetails in the workbench when displaying
the working directory or an unapplied patch.