nerge: use shelve tool to move working directory changes to a patch
When the merge wizard is launched from the Workbench, the shelve tool is opened
behind the merge wizard and I cannot figure out how to fix this.  In general Qt
does not want to open a QMainWindow from a modal dialog.  I'm having to do some
ugly hacks to get it to work this much.  This needs to be re-thought.