mq: more intelligent mixing of patch state and working directory state
A file could be added by the patch and modified by the working dir, we only
want to display that file once, as added.  Qref will combine them this way.
If a file is marked added or removed in the working directory, they should be
shown as A or R respectively.