fileview: unify diff/file/and annotate views
Use a single QSci instance for all three modes.  This fixes the margin text and
color issues in the manifest widget as well.  For whatever reason, you have to
rebuild the style each time the margin is refreshed, and it is also necessary to
set the font before setting the margin width else the margin size will be too
wide.