sync: simplify authentication dialog, rename to 'Security'
This dialog configures both https connection properties and user authentication.
It no longer offers the full feature set of the [auth] section as the complexity
is much too great for what is a rare requirement.  Also the previous logic was
defaulting to allowing authentication over http, which was transmitting your
password in plaintext if you happened to accidentally push over http.

It also disables the password entry if mercurial_keyring is enabled, since it
completely destroys the usefulness of mercurial_keyring if you go and save your
password in plaintext for zero benefit.

More work remains to be done, we need to add --insecure to the command line if
the hostname is found in [insecurehosts] and we are using https.