status: add a refresh toolbutton
Something very odd is happening here.  If I move the toolbutton to the left of
the line entry (essentially moving its addWidget call up two lines), the button
cannot be clicked when in the commit tool even though it is not disabled and
it works fine as part of the standalone status tool.