workbench: prepend repo shortname to window title
This helps better identify the entries in the taskbar of Windows.

Before this change, on e.g. Windows 7 you could typically only read
"TortoseHg Workben...". Now you can read -- for example --
"hg-main - Tortois...".

This helps to find the right entry in the taskbar of Windows if you
have multiple Workbench windows open.

Example titlebar text:

  Old:  TortoiseHg Workbench - C:\users\adi\hgrepos\hg-main
  New:  hg-main - TortoiseHg Workbench - C:\users\adi\hgrepos\hg-main

Compare with Mozilla Thunderbird, which shows the name of the selected
mail folder as the first component of the titlebar text.