hgemail: do not close main dialog when emails sent
I often do a test send to myself before sending a patch
to the "real" recipient, so I would like to decide myself
when the main email dialog window should be closed.

This is especially important when I've written an intro.
If the dialog is closed on send implicitly, I do have to
enter the intro text again for the real send.