status: Sort files in two passes: Sort by ascending path then sort by column
Until now the files in the status dialog could be sorted by checked state,
status, filename, extension or size. However, when sorting by extension, the
order of the files with the same extension was unspecified (same thing happened
when sorting by size or by checked state).

This patch fixes that problem by sorting in two passes:
1. Sort the list by file path (alphabetically, in ascending order)
2. Sort by the selected sort column