grep: include hg subrepositories in working folder scans
If the subrepo root path matches the filter, scan the entire repository.  This
does not allow file patterns to filter inside the subrepos themselves, it is an
all or nothing match.  Some of this code should be generalized and moved to
thgrepo so all tools that deal with workingctx can deal with subrepositories (on
the default branch).

Bonus fixes:

* Properly catch and report exceptions from the context scanning thread
* Do not hide exception messages by writing 'No matches found'