rejects: Show warning message when saving the file without having modified it
The rejects tool did not make it 100% clear that you are supposed to _edit_ the
file using its built-in editor. A user may assume that it only needs to mark all
conflicts as resolved and hit save, which would (magically?) resolve these
conflicts.

This patch shows a warning message when the user marks all rejected patch chunks
without having modified the file.