rupdate - Add dialog for Remote Update.
Requires rupdate: http://bitbucket.org/MrWerewolf/rupdate

Remote Update is adds the ability to update a remote repository to a
specified revision if the remote repository is accessed through ssh.

Note: The default TortoisePlink.exe will give you a "Not enough storage is
available to process this command" error when you try to run this. To get around
this, use another version of plink.  I use a copy of the TortoisePlink.exe from
TortoiseCVS.