fileview: Add find next/prev support to diff view
The way this has been done tries to follow the existing implementation for the
other file view modes (file and annotate). This means that the nextDiff and
prevDiff methods are unchanged.

The main difference between the diff mode and the other modes is that in order
to generate the diff markers we use a regular expression that matches the diff
"header" ("^@@ -[0-9]+,[0-9]+ \+[0-9]+,[0-9]+ @@$"), rather than callling
difflib.SequenceMatcher and parsing the resulting opcodes.