revdetails: Show modified subrepos on revision details panel
This change adds some basic support for showing modified subrepos on the
revision details.
Current limitations are:
- Due to a lack of subrepo specific icon, I am using the "hg" icon to represent
modified subrepos on the file list. This only makes sense for hg subrepos though
- It can only preview mercurial subrepos
- Behaviour for merge revisions is probably not perfect
- The "open..." link does not work. This is surprising since the code that is
used to generate the link (and the link contents) is exactly the same that is
used to generate the link on the status widget. I assume that the status and
revision details panels behave differently on this regard.