resolve: show info about the "local" and "other" revisions (closes #458)
This adds two csinfo widgets to the top of the resolve window, containing the
information about the local and the other revisions of a merge.

Note that the resolve window can be open even if we are in the middle of a merge
between two revisions (e.g. when merging local changes during and update). In
those cases the revision information is not trivial and for now it will simply
not be displayed.