fileview: don't create an empty subrepo unintentionally
When a subrepo is missing from the working directory, viewing the revision
history of any revision where said repository had been modified (initialized,
changed or removed) would create an empty subrepo in the working directory.

The problem was due to the fact that calling ctx.sub(wfile) always creates an
empty repository.

Many of the places where we were calling ctx.sub() were unneeded, since we only
needed to call it to get the current status of the repository, which does not
make sense if the repository is missing or if we are not viewing the working
directory status.

This patch could be further improved as follows:

- Avoid calling ctx.sub() for non hg subrepos, which could be done by
checking: ctx.substate.get(wfile, subrepo.nullstate)[2] != 'hg'
- Let the user create (clone?) missing subrepos.