filelistview: cleanup repo and context objects
The filelistmodel was given a repo even though it didn't use it anywhere,
meanwhile the filelistview had to reference the repo through the model.
Similarly, the current context was given directly to the model which was
then emitting the context to the view.  Instead, the context now goes through
the HgFileListView.