manifest: Show subrepo contents (files and subrepos) in the manifest treeview
This patch simply shows the files (and subrepos) that are contain within a
subrepo in the manifest tree view.
Files are only shown for mercurial subrepos which are present on the working
directory.
Currently the contents (or diff) of the items contained in subrepos is not shown
on the manifest file view.