tag: improve tag commit message when moving a tag
The original tag commit message when moving a tag was:
'Added tag %s for changeset %s'

This was wrong because the tag was not added, but moved. Instead this had been
changed to:
'Moved tag %s to changeset %s (from changeset %s)'

This is more accurate and also records a mention of the original changeset,
which, when viewed in the revision history is clickable.