manifest: properly show diffs of files modified within subrepos
This is a complex situation. If a second context is passed along to fileData
in order to explicitly compare two revisions (as it is usually the case),
then we must look for the subrepo that contains the file in the second context,
and for the subrepo revision that corresponds to that second context.

If no second revision is provided, the code normally tries to compare with the
first revision parent. Here we had two choices:
1.- Compare the subrepo files with _their_ parent revision
2.- Compare the subrepo files with the version in the top repository revision.

I've opted for #1, which makes more sense to me. Its drawback is that it will
show a diff that may not really exist in practice, since it may be a diff
between two non contiguous revisions.