workbench: add context menu to the tabbar.
The context menu actions are:
- Close
- Close other tabs

Also, if a tab has been closed before:
- Undo close tab

If a group of tabs has been closed before:
- Undo close other tabs

Note that when using "close other tabs" the tabbar is automatically hidden.
In order to get the "Undo close other tabs" option you must first open a second
repo so that the tab bar is shown again.

When clicking a non active tab, the tab will not be activated but the commands
in the context menu will apply to it nevertheless.
When clicking outside of any tab, the context menu will be shown and it will
apply to the current tab.