manifest: optimize subrepo search
In order to show the contents of subrepos in the manifest window, whenever a
file is selected in the manifest we must look for the deepest subrepo containing
the file.

Before this patch, the deepest subrepo search was performed by recursively
searching for subrepos from the top repo down, which could be an expensive
operation.

This patch adds a dictionary to the manifest model, which is generated once when
the model is created, containing path-context pairs (with the paths being the
keys to the dictionary). Then, when a file is selected in the manifest, we can
simply check this "subrepo info" dictionary, which is a much faster operation.