revdetails: always show selected file in navigate dialog
With multiple file selection, it's possible to have a DESELECTED current
item! Opening navigate dialog for the current item is confusing because other
actions operate on the selected items regardless the current item. This patch
aligns the File History and Compare file revisions actions with the other
menu entries behaviour.