repowidget: Change QGoto command into Apply/Unapply patch commands
QGoto is a confusing name for a command. Also, it has two possible uses or
results: It can apply or unapply a set of patches, depending on whether it is
used on an unapplied patch or on an applied patch.

This patch changes the QGoto command into "Apply patch (QGoto)" for unapplied patches,
and into an "Unapply patch (QGoto parent)" command for applied patches and qparent.

Note that "Unapply patch (QGoto)" is exactly the same as QGoto, while "Apply patch" is
the same as executing QGoto on the selected revision parent.

This patch has the drawback that selecting qparent the command name is still "Unapply patch", when it
should be "Unapply all patches". Actually we could simply hide the corresponding
command for all non qparent, since to unapply all patches you now just need to
unapply the lowest patch.