reporegistry: show non mercurial subrepos
This patch fixes most of the limitations with the way subrepos are shown on the
repo registry.

- It shows each subrepo type (hg, git, svn) with its own icon (git and svn icons
are based on icons found on the wikimedia site which have a Creative Commons
license)
- It overlays a "subrepo" icon on top of the subrepo icon type
This could have done by embedding the "subrepo" icon on top of each of the
subrepo icon types, but this solution seems more flexible and makes it easier
to change the subrepo marker icon in the future.

For non mercurial subrepos:
- It disables drag and drop
- It disables double clicking to open
- It hides those context menu actions that do not make sense for non mercurial
subrepos