reporegistry: do not load the subrepo list for repositories that cannot be open
This patch avoids trying to load the subrepo list if a repository that is
on the repo registry cannot be open. Whenever that happens, the error will be
silently ignored to replicate the behavior of thg before subrepo support was
added.

TODO: Add a "warning" overlay or some other sort of visual indicator to repos
that could not be load.