reporegistry: Add a warning indicator whenever a repo cannot be load
Currently the warning is only shown if thg is unable to load the list of
subrepos for a given repository.

Note that the warning icon is only set when the reporegistry is first load, so
any changes to the state of the repositories after the thg window has been open
will not be reflected in the repo registry status.

I am reusing the 'dialog-warning' icon. However maybe we should add an icon
named thg-warning to the scalable/status folder.