reporegistry: add "Refresh Registry list" action
Added a new command to the repository groups which allows refreshing the
repository registry list. This is useful in three scenarios:

- It allows to reload the repository list when there are multiple instances of
the workbench and the repo registry was modified in one of the other instances
(note that this only works when the other instance is closed)
- It allows manually reloading the list of subrepos that is shown on the
repository registry
- Building on the patch that shows a warning icon on repositories that cannot be
loaded, it allows to check whether a repository that was missing when the
workbench was loaded can now be accessed (or viceversa)

As a drawback, I added the command to the repostitory group, which may suggest
that only the corresponding group will be reload. However the command text
should make it clear that what is refreshed is the "Repository Registry", rather
than the group.