guess: improve renaming workflow
Do not require the user to select the files that will be looked for renames.
Instead, if no file is selected try to look for renames for all missing files.
Also, when no matches are selected the "Accept matches" button accepts all
matches (and it says so on the button text).