export: warn if a patch file already exists (closes #502)
The current behavior of export os to "append" the selected changes into any
existing patch file if the name of the file matches the name of the patch file
that would normally be created.

This patch adds a warning message that lets the user choose between the old
behaviour (which may have its uses) or a (IMHO) more sensible behavior in which
the existing patch is overwritten, or simply abort the operation.