reporegistry: Mark suprepos opened with "Open repository" with a subrepo icon
Note that for speed reasons we do not really check if the repository is a
subrepo. Instead we simply check whether it is _inside_ a mercurial repository,
in which case we assume that it is a subrepo.

If we really wanted to check whether the added repo is a subrepo we would need
to create a repo object for the container repo. Also, this may fail because the
selected revision on the container repo may be such that the repo that is being
open is not a subrepo in that revision, but it may well be one on another
revision. That is why it makes more sense to always mark repos inside repos as
"potential" subrepos.