fileview: show target revision first for modified subrepos
This is a port to stable of 2f730e81b07aef458beeeb8b227c30f949b46732, whose
description was as follows:

Until now when showing subrepo change information on the revision history
widget we were showing the source ("From:") revision first, followed by the
target ("To:") revision. If the size of the revision history widget was small
the user would only see the "From" information, which is less important thatn
the "To" information.

This patch changes the order in which that info is shown.
As a side effect, is improves the handling of invalid subrepos a bit, although
further improvements will be made on a later patch.