fileview: show proper message when subrepo is not missing but is invalid or missing revisions
This is a port to stable of 02f28fc92a81cd983f5f002c0f9f25a4c7e491b2, whose
description was as follows:

If a subrepo is missing, updating to a revision may create an "empty" subrepo
with no revisions. In those cases the message shown on the revision history was
wrong.