qimport: do not fail to qimport a revision when there is a patch name collision
qimport uses the default names for patches, whose format is "REVISION_NUMBER.diff".
Until this patch if you tried to qimport a revision when there was an existing
mq patch with the same target name, the qimport operation would fail, and there
was no obvious way to work around this problem.

It is easy to come into this problem. For instance, this happens if you qimport
a revision, unapply it, perform a new commit and then you try to qimport the new
commit.

This patch solves the problem by checking for name collisions before executing
the qimport command. If there are name collisions it will fall back to a slower
qimport method in which each revision that must be qimported will be imported
individually (starting from the newest), and for each of them thg will try to
find a unique patch name.