repowidget: stop treating dragged binary files as patches
Dragging a large binary file over any part of the workbench (even when the
workbench is sitting in the background behind other windows) will cause the
repowidget's dragEnterEvent to try to parse the file as a patch.  This will
cause the workbench (as well as the file explorer window from which the drag
originated) to hang for several seconds.