reporegistry: automatically update the repo registry when another thg process changes it
Until now different thg workbench processes were completely unaware of each other.
The result was that when several workbenches were open, only the last one managed
to set the state of its repo registry.

This patch addresses in two ways:
- It makes the repo registry immediately save its state
whenever a repo is added to or removed from, or dragged on the registry.
- It makes the repo registry monitor the repo registry settings file for changes.
Whenever a change is detected the repo registry is reload.

Problems and limitations:
- The QFileSystemWatcher object seems to get notified twice whenever the
settings file is modified. To avoid this problem and to limit of number of repo
updates, repo registry reloads due to settings file modification notifications
are delayed by 2 seconds. If multiple modification notifications are received
during that 2 second wait window they will trigger a single repo reistry reload.