repowidget: highlight the item corresponding to the active workbench tab
This patch sets the font of the repo registry item corresponding to the active
tab to bold.

This makes it very easy to spot the repo you are working with, in case you want
to use the repo registry item context menu to perform some operation on that
repository, or to view its list of subrepos, etc.