reporegistry: show paths relative to the common group path
This patch makes it possible to show only the part of the path that is not
common to all the repos on a repository group. This makes it much easier to
see the paths or the repos on the repo registry in those cases in which their
paths are very long.

This "Show short paths" mode can be enabled/disabled via a View menu checkbox.