mq: kill closeEvent on MQPatchesWidget
This gets called when the user closes the docking window, but
the MQPatchesWidget still stays araound anyway. If the window
is shown again, we had self.repo set to None, which is wrong
and caused a traceback

  Traceback (most recent call last):
    File "C:\Users\adi\hgrepos\thg-qt\tortoisehg\hgqt\mq.py", line 398, in onPatchSelected
      applied = set([p.name for p in self.repo.mq.applied])
  AttributeError: 'NoneType' object has no attribute 'mq'

as soon as a patch got selected again.

We simply don't need to do anything on closeEvent.