qtlib: define new class TaskWidget
Provides a default implementation of a function canswitch. If canswitch
returns True, the task widget "agrees" to switch away from it when a new
revision / patch is selected.

repowidget.RepoWidget.onRevisionClicked calls canswitch(). If it returns
False, no task tab switch occurs.

All task widgets are now also derived from qtlib.TaskWidget, so they get the
default canswitch() implementation, which always returns True.

manifestdialog.ManifestWidget overrides canswitch to return False. So the
decision if switching is allowed is moved into the implementation of that
widget.

qtlib.DemandWidget also got a canswitch function which returns True if
the widget wasn't yet created. If it's already created, it calls canswitch()
on the widget.

For the future, this design allows to deny a task tab switch dynamically,
that is, depending on some state of the task widget (not yet used).