sync: do not switch away from sync widget if target checkbox is checked
In sync widget, selecting a revision implicitly sets the target revision
if the target checkbox is checked.

Note that the user *still* can force the switch to the revision details widget
(by using for example the toolbar button). The switch is just no longer implicit
with the selection of a revision *iff* the target checkbox is checked.