cslist: workaround to avoid crash on Mac OS X with Qt 4.7.1 or later
A segmentation fault occurs on Mac OS X with Qt 4.7.1 or later
in the following cases:
- thg import patchfilename
- the user modifies the patchfilename in the src_combo
or
- thg import
- the user adds a patch filename in the src_combo (whatever the way:
Browse..., Browse Directory, Import from Clipboard)

After changing the sizeConstraint to SetMaximumSize for csvbox,
no more crash occurs. This change has no effect on the minimum size
of the csvbox as it is already limited by the size of the three
buttons above.