repotreeitem: use os.path.normcase() rather than str.lower() when comparing paths
The RepoItem.getRepoItem() method compares a given root path to the RepoItem
object root path. To do so it used to manually checking the os.name to see if
we are running on windows, in which case it "normalized" the paths by converting
them to lower case.

The correct solution for this problem is to use os.path.normcase(), which for
non Windows platforms returns the unmodified input, while for windows returns
the input string as lower case.