repowidget: Switch to the commit widget when selecting the topmost pacth
Now that the commit widget allows to perform qrefresh and qnew, it is preferable
to switch to the commit widget rather than to the mq widget when selecting the
topmost patch.

There is an existing issue which this patch does not fix, which is that if you
you run qimport while the working directory parent revision is selected, the
commit widget is _not_ selected. Instead the revision history widget will remain
selected.