manifest: improve subrepo view code
This patch contains several improvmenets to the code that is used to identify
the subrepo to which a file belongs to. The improvements are:

- It the name of one subrepo were a substring of another subrepo name, thg could
think that a file belonging to the subrepo with the longer pathname could belong
to the subrepo with the shorter pathname.
- There could be problems with paths with mixed file separators (i.e. '\\' and
'/').
The safest solution for this last problem would be to use os.path.normpath, but
it seems that paths are always stored in "unix" format internaly, so it seems
safe (and faster) to simply use "/" as the path separator when creating and
using the _subinfo dictionary.