filedata: do not crash when a subrepo .hgsubstate file is malformed (closes #783)
It seems that some (old) versions of mercurial had a rare bug that could result
in a malformed .hgsubstate file. In particular, it could result in
an .hgsubstate file where one or more of the lines referred to a revision id
that included a '+' character at the end, which is invalid.

When this happens, rather than crashing we will show a "WARNING" line on the
file data, but we will retrieve the information of the correct revision id.

Props to Mads Kiilerich for finding the reason for this problem.