workbench, reporegistry: open the cloned repository after successful clone
There was a bug in which when cloning a repository the cloned repository would
not be open. The repository _selected_ on the repo registry would be
open instead.

This changeset fixes that issue.

NOTE: This changeset changes the reporegistry.open method. That method checks
that the repository type is 'hg' before actually opening the selected repository.
Now this method may be used to open repos other than the selected one, repos
that do not have a reporegistry entry yet. In those cases we do not khow the
repotype yet. This has been solved by checking whether the .hg folder exists. If
it doesn't the repotype is set to 'unknown'.