commit: select QNew for the working dir and QRefresh for the topmost applied patch
This patch tries to fix some recent comments by Phil Currier regarding the
automatic switch to the commit widget when the topmost applied patch is selected.

The patch changes the behaviour of the commit widget to reduce the chances of
a user getting confused in the following two scenarios:

1.- When the user selects the working dir and the selected mode is "QRefresh":
This could lead to a user to refresh a patch by accident when in fact it wanted
to create a new patch.

2.- When the user selects the topmost patch while the "commit" button on the
commit widget is set to "QNew":
This could lead the user to believe that the selected patch has different
contents than expected.