repowidget: only show push context menu for outgoing changesets
This patch hides the push submenu of the repowidget context menu unless the user
clicks on an outgoing changeset. That is, in order to access the menu the user
must first execute the outgoing command and then right click on one of the
changesets.

The idea is to mirror the "bundle context menu" functionality, simplifying the
changeset context menu, which is already quite crowded.