run: fix aborting on displaying translated exception strings
With Mercurial translated output, this caused a decoding error to
be thrown, so the whole application would abort, but displaying
only the original exception.

The fix works for me for a lookup error browsing the manifest of
an empty repository; I'm not sure how to induce other unhandled
errors, so this may be not enough.