filedialogs: add context menu to "Compare file revisions" dialog
Note that I had only been able to test the Edit Local, View at revision and
revert commands, because I am unable to run the other commands when running
from source (the diff tool does not open).

However, I have the exact same problem with the File History dialog, and the
code is exactly the same, so it should work fine when not running from source.