settings, repowidget: add settings to select default widget and initial revision (closes #399)
Issue #399 proposes being able to let tortoisehg to select the working directory
when opening a repository.

This patch adds two new settings that not only allow configuring tortoisehg as
requested on issue #399, but also allow the user to select his preferred widget,
which can be any of the tortoisehg widgets (revision details, commit, manifest,
sync or search). This preferred widget is the one that will be selected by
default when opening a repository.

In addition, this patch adds a setting to select which revision should be
selected by default when opening a repository. It is possible to select the
current (i.e. the working directory parent) revision, the tip of the repository
or the working directory revision.

These two settings can be combined in interesting ways. For example it is
possible to make the commit widget the default while selecting the current
revision by default. It is also possible to make the working directory the
default and make the commit widget the default.

Note that when the working directory and the revision details widget are
selected by default, the commit widget will be selected instead.