export: show message after export, with a link to the export folder and patch
When a single revision is exported, the information message contains a link
to the export folder and a link to the patch file itself.

When multiple revisions are exported, the export count is shown plus a link
to the export folder.

Note that this does not check whether the export operation was successful, since
there is no easy way to get that information.