qtlib: HgRepoView.forward() must forward keyword parameters
The main side effect of this bug was that "push to revision" and
"push selected branch" where not working.

With the original version of the forward() method it was not really possible to
pass keyword parameters to the forwarded method because the forward() method
itself could receive a keyword parameter named "default". This was used in
several forward calls to "canExit" with "default=True".

The solution has been to add a default canExit() method to the TaskWidget and
DemandWidget classes, and directly call canExit where necessary.

Meanwhile the original forward() method has been modified to no longer accept a
"default" keyword argument, while passing any keyword arguments to the forwarded
method.