shellext: remove unused auto_buffer
The comment in the file that claims that std::vector<> provides no
guarantees about internal organization is wrong anyway.

The C++ standard guarantees that the elements of a std::vector are
consecutive in memory and we can get the address of the first element
with & and [], e.g. with

  std::vector<char> product_id(50, 0);

we can say

  &product_id[0]

for the address of the first element and it is safe.

So I'm mostly using vector anyway for such things.