workbench: do not reopen repos during startup until subrepos have been loaded into the registry (closes #1017)
During startup the workbench gets the list of repos (and subrepos) that were
open on the last session. It goes through the list, opening them one by one, and
if any is not found on the registry, it adds it automatically adds it to the
reporegistry default group.

Before this patch, the reload operation was performed before the subrepos had
been loaded into the registry. Thus if a subrepo was on the reopen list it would
never be found on the registry, and thus it would always be added to the default
group on startup.

This patch simply changes the order in which the settings are read. The "Save
open repositories on exit" setting is read after the "Show subrepos" setting has
been read and acted upon.