repotreeitem: Handle all exceptions that may happen during appendSubrepos (closes #1175)
We want to handle all exception that may happen while the reporegistry
is loading
the subrepos, because otherwise THG may not load at all.

Some expected exceptions are silently handled, marking the corresopnding repo
and/or subrepos as invalid. For all other exceptions we show a warning message
to the user, and we let him continue or abort, in which case the exception is
re-raised, so that we get the full exception error dialog.

Note that #1175 is marked as a duplicate of #995 on bitbucket, but these two
issues are totally unrelated.