reporegistry: allow dragging groups into groups
In order to still be able to drag groups to top level, dragging groups
behaves differently depending on whether you drag the group into another
group or into a repo inside a group.

Dragging a group into an repo inside a group will put the repo inside
the corresponding group (as could be expected).

Dragging a group into a group will place the group _at the same level_
as the target group (i.e. it will not drag the group _inside_ the
target group). This may be a bit counter-intuitive, but is the only
way I've found to still allow dragging a group to the top level.