commit, manifest: "edit" clean and untracked files when they are double clicked
Previosly double clicking on a clean or an untracked file on the status widget
would do nothing, while double clicking on one of those on the manifest widget
would try to diff it, resulting on a "No file changes" dialog being shown.

With this change the file will be open on the configured Visual Editor instead.