repoview: change unapplied patch drop indicator into a line
Up until now, when dragging unapplied patches to reorder them a rectangular
drop indicator appeared. This drop indicator appeared on top of the patch below
the mouse. This was confusing since it was not possible to tell whether the
patch that was being dragged would be dropped above or below the patch that was
marked with the rectangle.

This patch changes the drop indicator into a thin line that appears where the
patch will be dropped.