repowidget: capture commit widget output (related to #1379)
This path should capture errors generated by the commit widget and show them on
the info bar.

This is a simple fix for the issue where trying to commit with a dirty subrepo
on THG 2.2 silently fails. This is not a full fix for #1379 since ideally we
should let the user choose whether to commit the changes (using the --subrepos
flag) or not.