reporegistry: show proper short path when common group path ends with slash
A bug would happen if the common group path was a hard drive root, for example.
os.path.normpath() usually removes the trailing slash from a path, but it does
not if the path is a hard drive root (e.g. 'c:\') or a network root (e.g. '\\').

In those cases the short path would be wrong (the first letter would be missing).
This patch fixes that problem.