repomodel: initialize the branch colors in alphabetical order to make them more stable
Up until now the branch colors are a bit random. They change depending on which
revision is the working directory parent.

This patch makes the branch colors more stable by pre-initializing them in
alphabetical order. In addition the 'default' branch color is selected first,
so that it is always the same.