repowidget: Make double clicking an applied patch qgoto that patch (closes #1044)
The original behaviour was unapplying the double clicked patch (qgoto its parent).
This was the same behavior as the "Modify history/Unapply patch (QGoto parent)"
command on the context menu.

Note that the context menu command remains unchanged. This is desirable because
it makes it easy to unapply the lowest patch, without having to look for its
qparent.