sync: add support for the --mq option in the sync tool
This patch adds a combobox to the sync options dialog if the mq extension is
enabled. If one checks that combobox, the sync commands know that they should
work on the patch queue by setting the --mq option. If the mq extension is
disabled while the combobox is still checked, the stored mq option choice gets
discarded.

This patch does change only apply to push and pull, but not to incoming and
outgoing.