Add support for the projrc extension
This revision adds support for the projrc extension (see below for an
explanation of that the extension does). It basically modifies the settings
dialog to show the projrc configuration and settings, but only when the projrc
extension is enabled:

- It adds an extra"Projrc" configuration section which is only visible when the
projrc extension is enabled. In this section the user can set the servers from
which the projrc file will be pulled, the keys that will be pulled and whether
it must be prompted when the projrc file changes.

- It adds an extra, readonly panel to the settings dialog, showing the contents
of the projrc file (see below).

* What is the projrc extension:

The projrc extension (http://bitbucket.org/aragost/projrc) is meant to allow
setting a common "project" mercurial configuration which all users share. This
is especially useful in corporate environments where it is important to ensure
that all users share a common default configuration. For example, this
extension may be used to enable certain extensions or to setup certain hooks
for all users. It can also be used to subrepository subpaths, or to configure