repogistry: add cloned repo to the same group as the source repo
Up until now ane new repo is added to the default group of the repo registry.
When cloning a repo it is often desirable to have the source and the target
repos on the same group. With this patch clones will always be added to the
same group as the source repo.

If the source repo is not on the repo registry, the cloned repo will be added
to the default group instead.