workbench: add "Go to a specific revision" button to "edit toolbar"
I've created a new icon based on the "go-jump" and the "hg-log" icons. I find
that the go-jump icon looks too much like a "redo" icon and looks confusing in
this context.

As far as I know the go-jump icon is only used in the GoToQuickBar, but I don't
know where that is used. I wonder if we could remove it?